refactor: extractVvppをVvppManagerから切り出す
