+++ src/helpers/progressHelper.ts
